add jwt authentication to property endpoints

+++ api/endpoints/property.py
@@ -1,11 +1,29 @@
+import json, os, random, string, shutil
 from datetime import datetime
-from flask import Blueprint, jsonify, request
+from flask import Blueprint, current_app, jsonify, request
 from marshmallow import ValidationError
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_identity
+)
 from app import db
-from app.base.models import Property
+from app.base.image_handler import property_image_handler
+from app.base.models import Property, User
 from api.schema import property_schema, properties_schema
 
 property_endpoint = Blueprint("property_blueprint", __name__)
+
+
+def create_images_folder(username):
+    """
+    Generates a random string which will be used as a folder name for storing image files
+    of properties uploaded by user.
+    """
+    s = string.ascii_letters
+    output_str = "".join(random.choice(s) for i in range(10))
+    property_img_folder = f"property_images/{username}-property{output_str}"
+    os.mkdir(f"{current_app.root_path}/base/static/{property_img_folder}")
+    return property_img_folder
 
 
 @property_endpoint.route("/api/property/list")
@@ -31,61 +49,113 @@ def get_one_property(prop_id):
 
 
 @property_endpoint.route("/api/property/add", methods=["POST"])
+@jwt_required
 def add_property():
     """
     Creates a new property in the database with data from request body.
     """
-    data = request.get_json()
+    data = request.form
+    current_user = get_jwt_identity()
+    if "photos" not in data and request.files["photos"].filename == "":
+        return jsonify({"message": "You need to upload photos."})
     try:
+        image_files = request.files.getlist("photos")
+        folder_to_save_images = create_images_folder(current_user)
+        # List of image filenames to save to database
+        image_list = property_image_handler(image_files, folder_to_save_images)
+
+        # convert the python dictionary(image_files) to json string
+        image_list_to_json = json.dumps(image_list)
+        # Check if the request data matches the schema else raise ValidationError
         prop_data = property_schema.load(data)
         prop_to_add = Property(
             name=prop_data["name"],
             desc=prop_data["desc"],
             price=prop_data["price"],
             location=prop_data["location"],
-            photos=prop_data["photos"],
+            image_folder=folder_to_save_images,
+            photos=image_list_to_json,
             user_id=prop_data["user_id"],
         )
         db.session.add(prop_to_add)
         db.session.commit()
         return jsonify({"message": f"The property '{data['name']}' has been created."})
 
-    except ValidationError as err:
-        return jsonify({"ERROR": err.messages})
+    except ValidationError as err:  # Returns an error if a required field is missing
+        return jsonify({"error": err.messages})
 
 
-@property_endpoint.route("/api/property/update/<int:prop_id>", methods=["PUT"])
-def update_property(prop_id):
+@property_endpoint.route("/api/property/update", methods=["PUT"])
+@jwt_required
+def update_property():
     """
     Updates the details of the property in the database.
     """
-    data = request.get_json()
-    prop_to_update = Property.query.get(prop_id)
-    if not prop_to_update:
-        return jsonify({"error": f"The property with ID {prop_id} does not exist"})
+    data = request.form
+
+    user = get_jwt_identity()  # We get the username from token accessing this URL
+    owner = User.query.filter_by(
+        username=user
+    ).first()  # Then use the username from the token to query a user from db
+    prop_to_update = Property.query.get(data["id"])
+    if not owner:
+        return jsonify({"error": f"User not found"}), 403
+    elif not prop_to_update:
+        return jsonify({"message": "Invalid ID"})
+    elif prop_to_update.user_id != owner.id:
+        return jsonify({"message": "Action forbidden"}), 403
+
     try:
+        if "photos" not in data and request.files["photos"].filename == "":
+            return jsonify({"message": "You need to upload photos."})
+        # Check if the request data matches the schema else raise ValidationError
         prop_data = property_schema.load(data)
+        image_files = request.files.getlist("photos")
+        images_folder = prop_to_update.image_folder
+        # Returns a list containing folder where images are saved and filenames
+        image_list = property_image_handler(image_files, images_folder)
+
+        image_list_to_json = json.dumps(image_list)
+
         prop_to_update.name = prop_data["name"]
         prop_to_update.desc = prop_data["desc"]
         prop_to_update.price = prop_data["price"]
         prop_to_update.location = prop_data["location"]
-        prop_to_update.photos = prop_data["photos"]
+        prop_to_update.photos = image_list_to_json
         prop_to_update.date = datetime.utcnow()
-        prop_to_update.user_id = prop_data["user_id"]
+        prop_to_update.user_id = owner.id
         db.session.commit()
-        return jsonify({"message": "The details of the property has been updated."})
+        return (
+            jsonify({"message": "The details of the property has been updated."}),
+            201,
+        )
     except ValidationError as err:
         return jsonify(err.messages)
+    except KeyError as err:
+        return jsonify(err)
 
 
-@property_endpoint.route("/api/property/delete/<prop_id>", methods=["DELETE"])
-def delete_property(prop_id):
+@property_endpoint.route("/api/property/delete", methods=["DELETE"])
+@jwt_required
+def delete_property():
     """
     Deletes a property from database by querying using the ID from the url.
     """
-    prop_to_delete = Property.query.get(prop_id)
+    current_user = get_jwt_identity()
+    property_id = request.json["id"]
+    owner = User.query.filter_by(username=current_user).first()
+    if not owner:
+        return jsonify({"message": "User not found"}), 404
+    prop_to_delete = Property.query.get(property_id)
     if not prop_to_delete:
-        return jsonify({"error": f"The property of ID {prop_id} does not exist."})
+        return jsonify({"error": f"Invalid property ID."}), 404
+
+    if prop_to_delete.user_id != owner.id:
+        return jsonify({"message": "Action forbidden"}), 403
+    images_directory = os.path.join(
+        f"{current_app.root_path}/base/static/{prop_to_delete.image_folder}"
+    )
+    shutil.rmtree(images_directory)
     db.session.delete(prop_to_delete)
     db.session.commit()
     return jsonify(
